Refactored real time data handler to utils folder
