+++ app/utils/realtime_data_handler.py
